Split combine.cpp, part 2
